Interface funcionando e resto adaptado

+++ main/src/main/java/modelo/tabelaPaginas/EntradaTP.java
@@ -1,22 +1,22 @@
 package modelo.tabelaPaginas;
 
 public class EntradaTP {
-    private boolean modificacao;
-    private boolean presenca;
-    private boolean uso;
-    private int tempoUltimoUso;
-    private int numQuadro;
-	private int numPagina;
-    
-    public EntradaTP(boolean modificacao, boolean presenca, boolean uso, int tempoUltimoUso, int numQuadro, int numPagina) {
-    	this.modificacao = modificacao;
-    	this.presenca = presenca;
-    	this.uso = uso;
-    	this.tempoUltimoUso = tempoUltimoUso;
-    	this.numQuadro = numQuadro;
+	private boolean modificacao; // M: indica se a página foi modificada
+	private boolean presenca;    // P: indica se a página está na memória principal
+	private boolean uso;         // Bit de uso -> algoritmo do Relógio
+	private long tempoUltimoUso; // tempo do último acesso -> algoritmo LRU.
+	private int numQuadro;       // número do quadro na memória principal
+	private int numPagina;       // Número da página
+
+	public EntradaTP(boolean modificacao, boolean presenca, boolean uso, long tempoUltimoUso, int numQuadro, int numPagina) {
+		this.modificacao = modificacao;
+		this.presenca = presenca;
+		this.uso = uso;
+		this.tempoUltimoUso = tempoUltimoUso;
+		this.numQuadro = numQuadro;
 		this.numPagina = numPagina;
-    }
-    
+	}
+
 	public boolean getModificacao() {
 		return modificacao;
 	}
@@ -35,10 +35,10 @@ public class EntradaTP {
 	public void setUso(boolean uso) {
 		this.uso = uso;
 	}
-	public int getTempoUltimoUso() {
+	public long getTempoUltimoUso() {
 		return tempoUltimoUso;
 	}
-	public void setTempoUltimoUso(int tempoUltimoUso) {
+	public void setTempoUltimoUso(long tempoUltimoUso) {
 		this.tempoUltimoUso = tempoUltimoUso;
 	}
 	public int getNumQuadro() {
